+++ lib/pymedphys/_experimental/pseudonymisation/__init__.py
@@ -1,3 +1,4 @@
+# Copyright (C) 2026 Matthew Jennings
 # Copyright (C) 2020 Stuart Swerdloff, Simon Biggs
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -15,6 +16,7 @@ import functools
 import json
 import logging
 import pathlib
+import sys
 from os.path import abspath, dirname, isdir, isfile
 from os.path import join as pjoin
 
@@ -33,6 +35,14 @@ from . import strategy
 HERE = dirname(abspath(__file__))
 
 IDENTIFYING_UIDS_FILEPATH = pjoin(HERE, "identifying_uids.json")
+
+DEPRECATION_REASON = (
+    "experimental pseudonymisation is deprecated and will be removed in a "
+    "future release. It hashes UIDs and some numeric values without a secret "
+    "key, so anyone who holds the original data can re-link or recover them. "
+    "See https://docs.pymedphys.com/en/latest/users/background/"
+    "dicom-deidentification.html"
+)
 
 
 @functools.lru_cache()
@@ -67,6 +77,9 @@ def get_default_pseudonymisation_keywords():
 
 
 def anonymise_with_pseudo_cli(args):
+    # A DeprecationWarning is hidden by default, so tell command-line users.
+    print(f"Warning: {DEPRECATION_REASON}", file=sys.stderr)
+
     if args.delete_unknown_tags:
         handle_unknown_tags = True
     elif args.ignore_unknown_tags:
@@ -169,7 +182,7 @@ def pseudonymise(dicom_input, output_path=None):
         logging.error("Please submit issue to PyMedPhys")
         # but continue on, the data might not contain the offending keywords
         # and if it does... there will be some kind of error raised
-    keywords_to_leave_unchanged = list("PatientSex")
+    keywords_to_leave_unchanged = ["PatientSex"]
 
     if isinstance(dicom_input, pydicom.dataset.Dataset):
         pseudo_ds = anonymise_dataset(